# .md/API-CONTRACT.yaml
openapi: 3.0.3
info:
  title: Turma do Rola - Comary — Contrato de API
  version: 0.12.0
  description: >-
    Contrato publicado incrementalmente, endpoint a endpoint, assim que cada
    contrato fica estável (TASK.md Seção 1.7 / GUARDRAILS.md regra implícita
    de rastreabilidade de API) — não é preciso esperar a tarefa de Backend
    inteira terminar. Frontend/Mobile começam a consumir assim que um
    endpoint aparece aqui, mesmo que a implementação completa da tarefa
    ainda esteja em andamento. Mudar um contrato já publicado é permitido,
    mas nunca silenciosamente — toda mudança de contrato incompatível
    incrementa `info.version` e é registrada na seção "Changelog" no final
    deste arquivo.

    Duas famílias de endpoint neste sistema:
      1. **PostgREST autogerado pelo Supabase** sobre as views públicas
         curadas (`ranking_publico`, `presenca_mensal_publica`) — consumido
         **diretamente** pelo Frontend com a chave `anon` do Supabase, sem
         passar pela camada de API própria (SDD.md Seção 7.5, ADR-005). É o
         único caso em que o Frontend fala com o Supabase sem intermediação
         de um Route Handler — decisão já tomada no SDD.md, não uma exceção
         nova. Documentado aqui (paths desta seção) porque também é
         superfície pública real, exposta à internet.
      2. **Route Handlers do Next.js (App Router)**, publicados a partir de
         BE-04 em diante — toda escrita e toda leitura autenticada da área
...
servers:
...
components:
  securitySchemes:
    supabaseAnonKey:
      type: apiKey
      in: header
      name: apikey
      description: >-
        Chave pública (`anon`) do Supabase — único credential usado pelas
        páginas públicas (T02/T03). Nunca a `service role` (TASK.md Seção
        1.2/1.7, GUARDRAILS.md regra 7).
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessao_interna
      description: >-
        Cookie de sessão emitido por `POST /api/auth/login` (`httpOnly`,
        `Secure`, `SameSite=Strict`, TTL 8-12h — ADR-004, BE-04). Único
        mecanismo de autenticação da área interna (modelo binário, sem
        RBAC — SDD.md Seção 7.2): sessão válida = acesso total. Toda rota
        de escrita futura desta área (BE-06 em diante) deve declarar este
        esquema de segurança.
  responses:
    SessaoInvalida:
      description: >-
        Sem sessão válida (`middleware.ts`, BE-04, GUARDRAILS.md regra 17) —
        toda rota de escrita da área interna, exceto `/api/auth/login` e
        `/api/auth/logout`, retorna isto quando o cookie `sessao_interna`
        está ausente, expirado ou com assinatura inválida. Resposta
        reutilizável — toda rota de escrita futura (BE-06 em diante) deve
        referenciar `#/components/responses/SessaoInvalida` para o caso
        401, em vez de redefinir o formato.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroSessaoInvalida"
  schemas:
...
    LancarRodadaBody:
      type: object
      description: >-
        Corpo de `POST /api/rodadas` (BE-08, RF-02). `confirmar_duplicidade`:
        mesmo contrato já usado por `POST/PUT /api/atletas` (RF-01.5/BE-06) —
        quando já existe uma rodada `lancada` com a mesma `data` e este
        campo não vem `true`, a escrita é recusada com `409` (RF-02.8).
      properties:
        data:
          type: string
          format: date
          description: Data civil da rodada (AAAA-MM-DD).
        confirmar_duplicidade:
          type: boolean
          default: false
        participacoes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ParticipacaoRodadaBody"
          description: >-
            Uma entrada por atleta que participou (ou não) da rodada — cada
            `atleta_id` só pode aparecer uma vez (mesma restrição de
            `UNIQUE(rodada_id, atleta_id)` de `app.participacao_rodada`,
            BE-02, refletida aqui como erro `400`).
      required: [data, participacoes]
      additionalProperties: false
...
    RodadaResponse:
      type: object
      description: Resposta de sucesso (`201`) de `POST /api/rodadas` (BE-08).
      properties:
        id:
          type: string
          format: uuid
        data:
          type: string
          format: date
        status:
          type: string
          enum: [lancada, excluida]
          description: Sempre `"lancada"` na resposta deste endpoint (soft-delete, BE-02/BE-09).
        criado_em:
          type: string
          format: date-time
        participacoes:
          type: array
          items:
            $ref: "#/components/schemas/ParticipacaoRodadaResponse"
      required: [id, data, status, criado_em, participacoes]
      additionalProperties: false
    ErroValidacaoRodada:
      type: object
      description: >-
        Corpo de erro `400` de `POST /api/rodadas` — corpo malformado ou que
        falha `LancarRodadaBody` (inclui o bloqueio de RF-02.6: evento de
        jogo para atleta ausente, reportado como um item de `detalhes` com
        `path: ["participacoes", <índice>, "eventos"]`).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroDuplicidadeRodada:
      type: object
      description: >-
        Corpo de erro `409` de `POST /api/rodadas` (RF-02.8) — emitido
        quando já existe uma rodada `lancada` com a mesma `data` e a
        requisição não trouxe `confirmar_duplicidade: true`.
      properties:
        error:
          type: string
          example: "duplicidade"
        rodadas_duplicadas:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              data:
                type: string
                format: date
              status:
                type: string
                enum: [lancada, excluida]
            required: [id, data, status]
      required: [error, rodadas_duplicadas]
      additionalProperties: false
...
    LogAuditoriaItem:
      type: object
      description: >-
        Uma entrada de `app.log_auditoria` (BE-09, RF-04.4/RF-04.5) — grava
        toda correção/exclusão de rodada (BE-09) e toda anonimização de
        atleta (BE-07, ADR-011). NUNCA inclui campo de autor individual
        (RN-12/GUARDRAILS.md regra 18) — a própria tabela não tem essa
        coluna. `valores_antes`/`valores_depois` são objetos livres (jsonb)
        cujo formato depende de `tipo_evento`: para `"correcao"`, contêm
        `status`/`eventos`/`pontos_acumulados` (e `ajuste_aplicado` só em
        `valores_depois`) da participação corrigida; para `"estorno"`,
        contêm `status` da rodada e, em `valores_depois`,
        `atletas_afetados`/`pontos_revertidos`/`substituicoes_vinculadas`;
        para `"anonimizacao"` (BE-07), `valores_antes` contém SOMENTE
        marcadores redigidos (`"[REDACTED]"`), nunca o dado pessoal real.
      properties:
        id:
          type: string
          format: uuid
        rodada_id:
          type: string
          format: uuid
          nullable: true
          description: Preenchido para `tipo_evento` `correcao`/`estorno`; nulo para `anonimizacao`.
        atleta_id:
          type: string
          format: uuid
          nullable: true
          description: Preenchido para `tipo_evento` `correcao`/`anonimizacao`; nulo para `estorno` (afeta a rodada inteira).
        tipo_evento:
          type: string
          enum: [correcao, estorno, anonimizacao]
        ocorrido_em:
          type: string
          format: date-time
        valores_antes: {}
        valores_depois: {}
      required:
        [id, rodada_id, atleta_id, tipo_evento, ocorrido_em, valores_antes, valores_depois]
      additionalProperties: false
    ErroConsultaLogAuditoria:
      type: object
      description: >-
        Corpo de erro `400` de `GET /api/log-auditoria` — `limit` inválido
        (não numérico, ≤ 0, ou acima do teto máximo de 200).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    RestricaoBody:
      type: object
      description: >-
        Corpo de `POST /api/restricoes` e `PUT /api/restricoes/{id}` (BE-12,
        RF-05.5). Sem campo de "motivo" (`UX-SPEC.md` T10 — "sem campo de
        motivo obrigatório no PRD-TECNICO... fora de escopo desta release").
      properties:
        atleta_a_id:
          type: string
          format: uuid
        atleta_b_id:
          type: string
          format: uuid
          description: >-
            Deve ser diferente de `atleta_a_id` — `400` caso contrário
            (`restricao_obrigatoria_atletas_distintos_check` no banco,
            BE-02, replicado aqui como validação de API).
      required: [atleta_a_id, atleta_b_id]
      additionalProperties: false
    RestricaoObrigatoriaResponse:
      type: object
      description: >-
        Resposta de todos os endpoints de `/api/restricoes*` (BE-12).
        `atleta_a_nome`/`atleta_b_nome` (via `apelido_exibicao`, RN-06)
        espelham de propósito os mesmos nomes de campo do contrato
        `restricoes_conflitantes` do ADR-010 — `BE-11` consome
        `app.restricao_obrigatoria` para montar aquele contrato a partir da
        mesma tabela.
      properties:
        id:
          type: string
          format: uuid
        atleta_a_id:
          type: string
          format: uuid
        atleta_a_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) do atleta em `atleta_a_id`."
        atleta_b_id:
          type: string
          format: uuid
        atleta_b_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) do atleta em `atleta_b_id`."
        ativo:
          type: boolean
        desativado_em:
          type: string
          format: date-time
          nullable: true
          description: >-
            Preenchido na primeira desativação (RN-11) — nunca sobrescrito
            por chamadas repetidas de `POST .../desativar` (idempotência,
            preserva o registro histórico real); volta a `null` depois de
            `POST .../reativar`.
        criado_em:
          type: string
          format: date-time
      required:
        [
          id,
          atleta_a_id,
          atleta_a_nome,
          atleta_b_id,
          atleta_b_nome,
          ativo,
          desativado_em,
          criado_em,
        ]
      additionalProperties: false
    ErroValidacaoRestricao:
      type: object
      description: >-
        Corpo de erro `400` de `POST/PUT /api/restricoes*` — corpo
        malformado, `uuid` inválido, ou `atleta_a_id === atleta_b_id`.
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroAtletaReferenciadoNaoEncontrado:
      type: object
      description: >-
        Corpo de erro `404` de `POST/PUT /api/restricoes*` quando
        `atleta_a_id`/`atleta_b_id` não corresponde a nenhum `app.atleta`
        existente — devolvido pela API em vez de deixar a `FK` do banco
        (`references app.atleta (id)`, BE-02) estourar como erro genérico
        `23503`.
      properties:
        error:
          type: string
          example: "Atleta não encontrado."
        atleta_id:
          type: string
          format: uuid
          description: id do atleta que não foi encontrado (o primeiro dos dois, se ambos faltarem).
      required: [error, atleta_id]
      additionalProperties: false
...
    SubstituicaoBody:
      type: object
      description: >-
        Corpo de `POST /api/rodadas/{id}/substituicoes` (BE-13, RF-06.1).
        `time_id` precisa pertencer à rodada da URL (`404` caso contrário).
      properties:
        time_id:
          type: string
          format: uuid
        atleta_sai_id:
          type: string
          format: uuid
        atleta_entra_id:
          type: string
          format: uuid
          description: >-
            Precisa ser diferente de `atleta_sai_id` — critério de aceite
            literal de BE-13: "tentar usar o mesmo atleta em 'sai' e 'entra'
            é bloqueado com mensagem clara".
      required: [time_id, atleta_sai_id, atleta_entra_id]
      additionalProperties: false
    SubstituicaoResponse:
      type: object
      description: >-
        Uma substituição registrada (`app.substituicao`) — puro registro de
        fidelidade histórica (RF-06.1/RF-06.3), sem nenhum campo de autor
        individual (RN-12) e sem nenhum efeito sobre
        `app.lancamento_pontos`.
      properties:
        id:
          type: string
          format: uuid
        rodada_id:
          type: string
          format: uuid
        time_id:
          type: string
          format: uuid
        atleta_sai_id:
          type: string
          format: uuid
        atleta_sai_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) do atleta em `atleta_sai_id`."
        atleta_entra_id:
          type: string
          format: uuid
        atleta_entra_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) do atleta em `atleta_entra_id`."
        criado_em:
          type: string
          format: date-time
      required:
        [
          id,
          rodada_id,
          time_id,
          atleta_sai_id,
          atleta_sai_nome,
          atleta_entra_id,
          atleta_entra_nome,
          criado_em,
        ]
      additionalProperties: false
...
    ErroTimeNaoEncontradoNaRodada:
      type: object
      description: >-
        Corpo de erro `404` de `POST /api/rodadas/{id}/substituicoes` —
        `time_id` inexistente ou pertencente a outra rodada (não a `id` da
        URL).
      properties:
        error:
          type: string
          example: "Time não encontrado nesta rodada."
      required: [error]
      additionalProperties: false
    RodadaResumoItem:
      type: object
      description: >-
        Um item de `GET /api/rodadas` (BE-16, T06 do `UX-SPEC.md` —
        "Histórico de Rodadas"). Lacuna deixada por BE-08/BE-09/BE-10
        (nenhuma tarefa anterior cobria LEITURA de listagem de rodada — BE-08
        documentou a omissão explicitamente na época; BE-09/BE-10 cobriram
        só log de auditoria e preview de correção, nunca a listagem em si) —
        fechada nesta versão por decisão explícita do usuário ao identificar
        o bloqueio de `FE-06`. `presentes` é o único campo além dos já
        publicados por `RodadaResponse`/`RodadaExcluidaResponse`
        (`id`/`data`/`status`/`criado_em`), literal do wireframe T06
        ("19/09/2026 · 18 presentes").
      properties:
        id:
          type: string
          format: uuid
        data:
          type: string
          format: date
        status:
          type: string
          enum: [lancada, excluida]
          description: >-
            Rodada `excluida` (soft-delete) aparece normalmente na lista,
            com o status visível — NUNCA escondida silenciosamente (decisão
            de detalhe, mesmo racional já usado para `restricao_obrigatoria`
            desativada em `RestricaoObrigatoriaResponse`/T10, BE-12: o
            histórico permanece visível, quem decide o que fazer com uma
            rodada excluída na tela é o Frontend, não uma omissão do
            Backend).
        criado_em:
          type: string
          format: date-time
        presentes:
          type: integer
          description: >-
            Contagem de `participacao_rodada.status = "presente"` desta
            rodada (nunca inclui `lesionado`/`ausente` — mesma restrição já
            usada por `RankingPublicoItem.presencas`, BE-03/BLOCKER-005,
            embora aqui sem o filtro de `status = "lancada"` da rodada, já
            que rodada `excluida` também aparece nesta listagem).
      required: [id, data, status, criado_em, presentes]
      additionalProperties: false
    ErroConsultaRodadas:
      type: object
      description: >-
        Corpo de erro `400` de `GET /api/rodadas` — `limit` inválido (não
        numérico, ≤ 0, ou acima do teto máximo de 200), mesmo formato de
        `ErroConsultaLogAuditoria` (BE-09).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
...
paths:
...
  /api/rodadas:
    get:
      tags: [interno]
      summary: Lista rodadas em ordem cronológica decrescente (BE-16, T06 do UX-SPEC.md)
      description: >-
        Lacuna deixada por BE-08/BE-09/BE-10 (nenhuma delas cobria LEITURA de
        listagem de rodada — BE-08 documentou a omissão explicitamente na
        época) — fechada nesta versão por decisão explícita do usuário ao
        identificar o bloqueio de `FE-06`. Lista `app.rodada` ordenada por
        `data desc` e, em caso de empate (duas rodadas na mesma data civil,
        cenário válido via `confirmar_duplicidade`, RF-02.8), por
        `criado_em desc` como desempate. NUNCA filtra por `status` — rodada
        `excluida` (soft-delete) aparece normalmente, com o status visível
        (ver `RodadaResumoItem`). Leitura interna protegida por sessão mesmo
        sendo `GET` (mesmo racional já usado por `GET /api/log-auditoria`,
        BE-09) — `middleware.ts`, `INTERNAL_READ_PROTECTED_PREFIXES`
        (`/api/rodadas` já estava na lista desde BE-13, nenhuma mudança
        necessária nesta tarefa).
      operationId: getRodadas
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: >-
            Teto de linhas retornadas (decisão de detalhe, TASK.md Seção
            1.0 — não exigido literalmente pelo critério de aceite; mesmo
            racional/mesmos valores de `GET /api/log-auditoria`, BE-09:
            `app.rodada` cresce indefinidamente ao longo das temporadas, sem
            expurgo automático).
      responses:
        "200":
          description: Lista ordenada do mais recente ao mais antigo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RodadaResumoItem"
        "400":
          description: "`limit` inválido (não numérico, ≤ 0, ou acima de 200)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroConsultaRodadas"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Lança uma rodada — presença, eventos de jogo e cálculo automático de pontos (BE-08, RF-02)
      description: >-
        Aciona a função PL/pgSQL `app.lancar_rodada`, que roda dentro de uma
        única transação (TASK.md Seção 1.2): grava `app.rodada`, uma
        `app.participacao_rodada` por atleta, `app.evento_jogo` por evento
        de gol/cartão e exatamente um `app.lancamento_pontos` por atleta
        (ledger append-only, ADR-006) com o saldo já calculado. Uma
        exceção em qualquer ponto do processamento reverte 100% do que já
        tinha sido gravado nesta chamada (RNF-10 — nunca estado parcial
        visível). Valor de pontos de cada evento sempre lido do vigente em
        `app.configuracao_pontuacao` na data da rodada (RN-05: presença
        `+2`, ausência `0`, lesionado pontua como presença, gol `+3`,
        cartão amarelo `-1`, cartão vermelho `-3`) — nunca hardcoded.
        Bloqueia com `400` se qualquer participação com `status: "ausente"`
        trouxer `eventos` não vazio (RF-02.6) — validado na própria camada
        de API (`zod`) e repetido como defesa em profundidade estrutural
        dentro da função PL/pgSQL. Alerta de rodada duplicada (RF-02.8) —
        ver `409` abaixo; rodada com `status: "excluida"` (soft-delete,
        BE-02) na mesma `data` nunca dispara esse alerta. Substituições
        (RF-06) não fazem parte deste endpoint — registradas depois, a
        partir de times já definidos (BE-13, fora do escopo de BE-08).
      operationId: postRodadas
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LancarRodadaBody"
      responses:
        "201":
          description: >-
            Rodada lançada. Cada item de `participacoes` inclui
            `pontos_delta` — o valor já efetivamente gravado em
            `app.lancamento_pontos` para aquele atleta nesta rodada (soma de
            presença/ausência + todos os eventos de jogo do atleta).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RodadaResponse"
        "400":
          description: >-
            Corpo malformado/inválido — inclui o bloqueio de RF-02.6 (evento
            de jogo para atleta ausente) como um item de `detalhes` com
            `path: ["participacoes", <índice>, "eventos"]`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoRodada"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "409":
          description: >-
            Já existe uma rodada `lancada` (nunca `excluida`) com a mesma
            `data` e a requisição não trouxe `confirmar_duplicidade: true`
            (RF-02.8) — nenhuma rodada foi criada. Reenviar a mesma
            requisição com `confirmar_duplicidade: true` para confirmar e
            lançar mesmo assim.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDuplicidadeRodada"
...
  /api/log-auditoria:
    get:
      tags: [interno]
      summary: Consulta o log de auditoria de correções/exclusões e anonimizações (BE-09, RF-04.5)
      description: >-
        Lista `app.log_auditoria` ordenado do mais recente ao mais antigo
        (`ocorrido_em desc`, RF-04.5) — grava toda correção/exclusão de
        rodada (BE-09) e toda anonimização de atleta (BE-07, ADR-011).
        NUNCA inclui campo de autor individual (RN-12/GUARDRAILS.md regra
        18) — a própria tabela não tem essa coluna. Leitura interna
        protegida por sessão mesmo sendo `GET` (RF-07.1 lista "histórico/
        correção" entre as ações que exigem a senha interna) —
        `middleware.ts`, `INTERNAL_READ_PROTECTED_PREFIXES`, mesmo padrão
        já usado por `GET /api/atletas` (BE-06).
      operationId: getLogAuditoria
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: >-
            Teto de linhas retornadas (decisão de detalhe, TASK.md Seção
            1.0 — não exigido literalmente pelo critério de aceite; evita
            crescimento não controlado da resposta, já que
            `app.log_auditoria` é retido indefinidamente sem expurgo
            automático, RNF-06).
      responses:
        "200":
          description: Lista ordenada do mais recente ao mais antigo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogAuditoriaItem"
        "400":
          description: "`limit` inválido (não numérico, ≤ 0, ou acima de 200)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroConsultaLogAuditoria"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
  /api/restricoes:
    get:
      tags: [interno]
      summary: Lista restrições obrigatórias, ativas e desativadas (BE-12, RF-05.5)
      description: >-
        Exige sessão válida mesmo sendo leitura (mesmo racional de
        `GET /api/log-auditoria` — feature exclusiva da área interna, T10 do
        `UX-SPEC.md`; `middleware.ts`, `INTERNAL_READ_PROTECTED_PREFIXES`).
        Devolve TODAS as restrições, ativas e desativadas (RN-11 — histórico
        nunca some da lista, apenas `ativo`/`desativado_em` mudam),
        ordenadas com as ativas primeiro, mais recente para mais antiga
        dentro de cada grupo (decisão de detalhe, não exigida literalmente
        pelo critério de aceite). Cada item já vem com
        `atleta_a_nome`/`atleta_b_nome` (RN-06) resolvidos, evitando
        chamadas extras no Frontend (T10).
      operationId: getRestricoes
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Lista de restrições.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestricaoObrigatoriaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Cadastra uma restrição obrigatória entre dois atletas (BE-12, RF-05.5)
      description: >-
        Cria com `ativo=true`/`desativado_em=null`. Recusa com `400` se
        `atleta_a_id === atleta_b_id`; com `404` se algum dos dois ids não
        corresponder a um atleta existente. Sem verificação de duplicidade
        de par já cadastrado (decisão de detalhe, TASK.md Seção 1.0 — não
        exigida por RF-05.5/RN-11, diferente de RF-01.5 para
        `nome_completo` de atleta; um par duplicado é inofensivo para o
        grafo de conflito do ADR-010, aresta repetida não muda o resultado
        da checagem por componente conexo). RN-12: qualquer sessão válida
        pode chamar este endpoint — não há hierarquia nem campo de autor
        individual em nenhum lugar deste contrato.
      operationId: postRestricoes
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestricaoBody"
      responses:
        "201":
          description: Restrição criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestricaoObrigatoriaResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoRestricao"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: "`atleta_a_id` ou `atleta_b_id` não corresponde a nenhum atleta existente."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaReferenciadoNaoEncontrado"
...
  /api/rodadas/{id}/substituicoes:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: id da rodada (`app.rodada`) a que as substituições pertencem.
    get:
      tags: [interno]
      summary: Lista as substituições registradas de uma rodada (BE-13, RF-06.2)
      description: >-
        Exige sessão válida mesmo sendo leitura — não devolve dado pessoal
        sensível, mas é feature exclusiva da área interna (T11 do
        `UX-SPEC.md`, "Substituições registradas"), mesmo racional já usado
        para `GET /api/restricoes` (BE-12); `middleware.ts`,
        `INTERNAL_READ_PROTECTED_PREFIXES` (`/api/rodadas` adicionado nesta
        tarefa). Sem limite de quantidade (RF-06.2) — devolve todas,
        ordenadas da mais antiga para a mais recente.
      operationId: getSubstituicoesRodada
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Lista de substituições desta rodada (vazia se nenhuma foi registrada).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubstituicaoResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Registra uma substituição vinculada a uma rodada/time (BE-13, RF-06.1)
      description: >-
        Registro puro de fidelidade histórica (RF-06.1) — nunca gera
        pontuação própria (RF-06.3, critério de aceite literal: "registrar
        substituição não altera saldo de nenhum atleta"; nenhuma escrita em
        `app.lancamento_pontos`). Sem limite de quantidade por rodada
        (RF-06.2, critério de aceite literal: "múltiplas substituições na
        mesma rodada sem limite"). Recusa com `400` se
        `atleta_sai_id === atleta_entra_id` (critério de aceite literal:
        "tentar usar o mesmo atleta em 'sai' e 'entra' é bloqueado com
        mensagem clara" — mesmo reforço estrutural já existente no banco,
        constraint `substituicao_atletas_distintos_check`, BE-02). RN-12:
        qualquer sessão válida pode chamar este endpoint — não há
        hierarquia nem campo de autor individual em nenhum lugar deste
        contrato. Depende de `POST /api/rodadas/{id}/times` já ter
        persistido `app.time` para esta rodada (escopo ampliado desta mesma
        execução de BE-13) — sem isso, nenhum `time_id` válido existiria
        para `time_id` referenciar.
      operationId: postSubstituicaoRodada
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubstituicaoBody"
      responses:
        "201":
          description: Substituição registrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubstituicaoResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoSubstituicao"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: >-
            `time_id` inexistente ou pertencente a outra rodada
            (`ErroTimeNaoEncontradoNaRodada`) ou `atleta_sai_id`/
            `atleta_entra_id` não corresponde a nenhum atleta existente
            (`ErroAtletaReferenciadoNaoEncontrado`, já publicado por
            BE-12).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErroTimeNaoEncontradoNaRodada"
                  - $ref: "#/components/schemas/ErroAtletaReferenciadoNaoEncontrado"
# Changelog — toda mudança de contrato incompatível (ou, como neste caso,
# uma mudança aditiva que o SDD.md pede para versionar explicitamente,
# SDD.md Seção 5.1) incrementa `info.version` acima e ganha uma entrada
# aqui (convenção referenciada em `info.description`; `x-changelog` é
# extensão válida do OpenAPI 3.x para conteúdo fora do schema formal).
...